Add uploadFileOnLink and move uploadFileOn... resolvers to appropriate modules (#3597)
* wip

* Move uploadFileOnLink to link module.

* Move uploadFileOnReference to reference module.

* Move uploadFileOnVisual to visual module.
