+++ src/domain/common/reference/reference.module.ts
@@ -5,11 +5,15 @@ import { AuthorizationPolicyModule } from '../authorization-policy/authorization
 import { Reference } from './reference.entity';
 import { ReferenceResolverMutations } from './reference.resolver.mutations';
 import { ReferenceService } from './reference.service';
+import { StorageBucketModule } from '@domain/storage/storage-bucket/storage.bucket.module';
+import { DocumentModule } from '@domain/storage/document/document.module';
 
 @Module({
   imports: [
     AuthorizationPolicyModule,
     AuthorizationModule,
+    StorageBucketModule,
+    DocumentModule,
     TypeOrmModule.forFeature([Reference]),
   ],
   providers: [ReferenceResolverMutations, ReferenceService],
