+++ src/domain/common/visual/visual.module.ts
@@ -6,11 +6,15 @@ import { Visual } from './visual.entity';
 import { VisualResolverMutations } from './visual.resolver.mutations';
 import { VisualService } from './visual.service';
 import { VisualAuthorizationService } from './visual.service.authorization';
+import { DocumentModule } from '@domain/storage/document/document.module';
+import { StorageBucketModule } from '@domain/storage/storage-bucket/storage.bucket.module';
 
 @Module({
   imports: [
     AuthorizationPolicyModule,
     AuthorizationModule,
+    DocumentModule,
+    StorageBucketModule,
     TypeOrmModule.forFeature([Visual]),
   ],
   providers: [
