+++ src/domain/storage/storage-bucket/dto/storage.bucket.dto.upload.file.on.link.ts
@@ -1,0 +1,7 @@
+import { InputType, Field } from '@nestjs/graphql';
+
+@InputType()
+export class StorageBucketUploadFileOnLinkInput {
+  @Field({ nullable: false })
+  linkID!: string;
+}
